+++ iOS/G3MiOSSDK/Commons/Rendererers/PlanetRenderer.hpp
@@ -469,7 +469,8 @@ public:
     return _verticalExaggeration;
   }
   
-  void setChangedRendererInfoListener(ChangedRendererInfoListener* changedInfoListener, const int rendererIdentifier);
+  void setChangedRendererInfoListener(ChangedRendererInfoListener* changedInfoListener,
+                                      const size_t rendererIdentifier);
   
 };
